Rename core to raw
