+++ examples/polkadot.rs
@@ -47,7 +47,7 @@ struct Header {
 fn main() {
     async_std::task::block_on(async move {
         let transport = jsonrpsee::transport::ws::WsTransportClient::new("wss://kusama-rpc.polkadot.io").await.unwrap();
-        let mut raw_client = jsonrpsee::core::RawClient::new(transport);
+        let mut raw_client = jsonrpsee::raw::RawClient::new(transport);
         let v = System::system_name(&mut raw_client).await.unwrap();
         println!("{:?}", v);
