+++ src/raw/server/notification.rs
@@ -25,7 +25,7 @@
 // DEALINGS IN THE SOFTWARE.
 
 use crate::common;
-use crate::core::server::Params;
+use crate::raw::server::Params;
 use core::fmt;
 
 /// Notification received on a server.
